backend/pyproject.toml
[tool.poetry]
name = "karaoke-api"
version = "0.1.0"
description = "Backend API for Instruments Karaoke"
authors = ["vitor-amartins <vmam@cin.ufpe.br>"]
readme = "README.md"
packages = [{include = "karaoke_api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
spleeter = "^2.4.0"
pytube = "^15.0.0"
Unidecode = "^1.3.6"
pydub = "^0.25.1"
youtube-search-python = "1.6.6"
pyacoustid = "^1.2.2"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
